zavrseno zatvaranje modala, postovanje svih tipova postova i relodovanje feed page-a nakon postovanja
